fix(compositing): MSVC narrowing-conversion in as_ndarray strides
std::vector<ssize_t>'s brace-init list disallows narrowing
conversions per the standard; cv::Mat::step returns size_t, and
size_t -> ssize_t (now typedef SSIZE_T on MSVC, see main-repo
common.hpp) is a real narrowing MSVC correctly rejects even though
GCC/Clang let this exact pattern through. Explicit static_cast on the
step[] elements makes the conversion happen before entering the
initializer list, which is allowed.

Found via Image-Toolkit PR #647's Windows build dry-run CI.

+++ base/src/compositing.cpp
@@ -36,10 +36,10 @@ inline py::array_t<uint8_t> as_ndarray(const cv::Mat& mat) {
     std::vector<ssize_t> strides;
     if (mat.channels() == 3) {
         shape = {mat.rows, mat.cols, 3};
-        strides = {mat.step[0], mat.step[1], 1};
+        strides = {static_cast<ssize_t>(mat.step[0]), static_cast<ssize_t>(mat.step[1]), 1};
     } else {
         shape = {mat.rows, mat.cols};
-        strides = {mat.step[0], 1};
+        strides = {static_cast<ssize_t>(mat.step[0]), 1};
     }
     auto result = py::array_t<uint8_t>(shape, strides);
     auto req = result.request();
